Rename the dadabase config files
