+++ package.json
@@ -4,8 +4,10 @@
   "description": "Projeto desenvolvido para gerenciar perguntas anônimas de usários direcionadas ao ADM da pagina. Desenvolvido durante semana NLW Rocketseat - Discover",
   "main": "src/server.js",
   "scripts": {
-    "start": "nodemon .",
-    "db-init": "node src/db/init.js"
+    "start": "node  src/server.js",
+    "start-dev": "nodemon .",
+    "db-init-postgres": "node src/db/init-postgres.js",
+    "db-init-sqlite": "node src/db/init-sqlite.js"
   },
   "Repositor": {
     "type": "git",
