--- src/db/config.js
@@ -1,7 +1,0 @@
-const sqlite3 = require("sqlite3")
-const { open } = require("sqlite")
-
-module.exports = () => open({
-    filename: './src/db/rocketq.lite',
-    driver: sqlite3.Database
-})
